@@ -11,10 +11,20 @@
 	<description>Easily generate docs and playground for your RESTful API</description>
 	<url>http://jsondoc.org</url>
 
+	<properties>
+		<java.version>10</java.version>
+		<maven.compiler.source>10</maven.compiler.source>
+		<maven.compiler.target>10</maven.compiler.target>
+		<spring.boot.version>2.0.5.RELEASE</spring.boot.version>
+		<spring.version>5.1.1.RELEASE</spring.version>
+		<org.reflections.version>0.9.10</org.reflections.version>
+		<log4j.api.version>1.7.25</log4j.api.version>
+	</properties>
+
 	<developers>
 		<developer>
-			<name>Fabio Maffioletti</name>
-			<email>fabio.maffioletti@gmail.com</email>
+			<name>Wavely IO</name>
+			<email>bug@wavely.io</email>
 		</developer>
 	</developers>
 
