@@ -8,27 +8,25 @@ import org.jsondoc.core.pojo.ApiObjectFieldDoc;
 import org.jsondoc.core.scanner.DefaultJSONDocScanner;
 import org.jsondoc.core.scanner.JSONDocScanner;
 import org.jsondoc.core.util.pojo.HibernateValidatorPojo;
-import org.junit.Assert;
-import org.junit.Test;
-
-import com.google.common.collect.Sets;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class JSONDocApiObjectBuilderTest {
 	
-	JSONDocScanner jsondocScanner = new DefaultJSONDocScanner();
+	private JSONDocScanner jsondocScanner = new DefaultJSONDocScanner();
 	
 	@Test
 	public void testApiObjectDocWithHibernateValidator() {
-		Set<ApiObjectDoc> apiObjectDocs = jsondocScanner.getApiObjectDocs(Sets.<Class<?>>newHashSet(HibernateValidatorPojo.class));
+		Set<ApiObjectDoc> apiObjectDocs = jsondocScanner.getApiObjectDocs(Set.of(HibernateValidatorPojo.class));
 		Iterator<ApiObjectDoc> iterator = apiObjectDocs.iterator();
 		ApiObjectDoc next = iterator.next();
 		Set<ApiObjectFieldDoc> fields = next.getFields();
 		for (ApiObjectFieldDoc apiObjectFieldDoc : fields) {
 			if(apiObjectFieldDoc.getName().equals("id")) {
 				Iterator<String> formats = apiObjectFieldDoc.getFormat().iterator();
-				Assert.assertEquals("a not empty id", formats.next());
-				Assert.assertEquals("length must be between 2 and 2147483647", formats.next());
-				Assert.assertEquals("must be less than or equal to 9", formats.next());
+				Assertions.assertEquals("a not empty id", formats.next());
+				Assertions.assertEquals("length must be between 2 and 2147483647", formats.next());
+				Assertions.assertEquals("must be less than or equal to 9", formats.next());
 			}
 		}
 	}
