@@ -9,14 +9,14 @@ import org.jsondoc.core.pojo.ApiVisibility;
 import org.jsondoc.core.pojo.JSONDoc.MethodDisplay;
 import org.jsondoc.core.scanner.DefaultJSONDocScanner;
 import org.jsondoc.core.scanner.JSONDocScanner;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
-import com.google.common.collect.Sets;
+import java.util.Set;
 
 public class JSONDocApiVisibilityBuilderTest {
 	
-	JSONDocScanner jsondocScanner = new DefaultJSONDocScanner();
+	private JSONDocScanner jsondocScanner = new DefaultJSONDocScanner();
 	
 	@Api(name = "test-type-level-visibility-and-stage", description = "Test type level visibility and stage attributes", visibility = ApiVisibility.PUBLIC, stage = ApiStage.BETA)
 	private class Controller {
@@ -44,29 +44,29 @@ public class JSONDocApiVisibilityBuilderTest {
 	
 	@Test
 	public void testApiVisibility() {
-		ApiDoc apiDoc = jsondocScanner.getApiDocs(Sets.<Class<?>> newHashSet(Controller.class), MethodDisplay.URI).iterator().next();
-		Assert.assertEquals(ApiVisibility.PUBLIC, apiDoc.getVisibility());
-		Assert.assertEquals(ApiStage.BETA, apiDoc.getStage());
+		ApiDoc apiDoc = jsondocScanner.getApiDocs(Set.of(Controller.class), MethodDisplay.URI).iterator().next();
+        Assertions.assertEquals(ApiVisibility.PUBLIC, apiDoc.getVisibility());
+        Assertions.assertEquals(ApiStage.BETA, apiDoc.getStage());
 		
 		for (ApiMethodDoc apiMethodDoc : apiDoc.getMethods()) {
 			if(apiMethodDoc.getPath().contains("/inherit")) {
-				Assert.assertEquals(ApiVisibility.PUBLIC, apiMethodDoc.getVisibility());
-				Assert.assertEquals(ApiStage.BETA, apiMethodDoc.getStage());
+                Assertions.assertEquals(ApiVisibility.PUBLIC, apiMethodDoc.getVisibility());
+                Assertions.assertEquals(ApiStage.BETA, apiMethodDoc.getStage());
 			}
 			if(apiMethodDoc.getPath().contains("/override")) {
-				Assert.assertEquals(ApiVisibility.PRIVATE, apiMethodDoc.getVisibility());
-				Assert.assertEquals(ApiStage.GA, apiMethodDoc.getStage());
+                Assertions.assertEquals(ApiVisibility.PRIVATE, apiMethodDoc.getVisibility());
+                Assertions.assertEquals(ApiStage.GA, apiMethodDoc.getStage());
 			}
 		}
 		
-		apiDoc = jsondocScanner.getApiDocs(Sets.<Class<?>> newHashSet(Controller2.class), MethodDisplay.URI).iterator().next();
-		Assert.assertEquals(ApiVisibility.UNDEFINED, apiDoc.getVisibility());
-		Assert.assertEquals(ApiStage.UNDEFINED, apiDoc.getStage());
+		apiDoc = jsondocScanner.getApiDocs(Set.of(Controller2.class), MethodDisplay.URI).iterator().next();
+        Assertions.assertEquals(ApiVisibility.UNDEFINED, apiDoc.getVisibility());
+        Assertions.assertEquals(ApiStage.UNDEFINED, apiDoc.getStage());
 		
 		for (ApiMethodDoc apiMethodDoc : apiDoc.getMethods()) {
 			if(apiMethodDoc.getPath().contains("/only-method")) {
-				Assert.assertEquals(ApiVisibility.PRIVATE, apiMethodDoc.getVisibility());
-				Assert.assertEquals(ApiStage.DEPRECATED, apiMethodDoc.getStage());
+                Assertions.assertEquals(ApiVisibility.PRIVATE, apiMethodDoc.getVisibility());
+                Assertions.assertEquals(ApiStage.DEPRECATED, apiMethodDoc.getStage());
 			}
 		}
 		
