@@ -1,57 +1,55 @@
 package org.jsondoc.core.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Sets;
-import org.hamcrest.CoreMatchers;
 import org.jsondoc.core.pojo.JSONDocTemplate;
 import org.jsondoc.core.util.pojo.ClassWithConstant;
 import org.jsondoc.core.util.pojo.TemplateObject;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.util.Assert;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.jsondoc.core.util.pojo.ClassWithConstant.THIS_IS_A_CONSTANT;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class JSONDocTemplateBuilderTest {
 
 	@Test
-	public void testTemplate() throws IOException, IllegalArgumentException, IllegalAccessException, InstantiationException {
+	public void testTemplate() throws IOException, IllegalArgumentException {
 		ObjectMapper mapper = new ObjectMapper();
-		Set<Class<?>> classes = Sets.<Class<?>>newHashSet(TemplateObject.class);
+		Set<Class<?>> classes = Set.of(TemplateObject.class);
 		
 		Map<String, Object> template = JSONDocTemplateBuilder.build(TemplateObject.class, classes);
 
-		Assert.assertEquals(0, template.get("my_id"));
-		Assert.assertEquals(0, template.get("idint"));
-		Assert.assertEquals(0, template.get("idlong"));
-		Assert.assertEquals("", template.get("name"));
-		Assert.assertEquals("", template.get("gender"));
-		Assert.assertEquals(true, template.get("bool"));
-		Assert.assertEquals(new ArrayList(), template.get("intarrarr"));
-		Assert.assertEquals(new JSONDocTemplate(), template.get("sub_obj"));
-		Assert.assertEquals(new ArrayList(), template.get("untypedlist"));
-		Assert.assertEquals(new ArrayList(), template.get("subsubobjarr"));
-		Assert.assertEquals(new ArrayList(), template.get("stringlist"));
-		Assert.assertEquals(new ArrayList(), template.get("stringarrarr"));
-		Assert.assertEquals(new ArrayList(), template.get("integerarr"));
-		Assert.assertEquals(new ArrayList(), template.get("stringarr"));
-		Assert.assertEquals(new ArrayList(), template.get("intarr"));
-		Assert.assertEquals(new ArrayList(), template.get("subobjlist"));
-		Assert.assertEquals(new ArrayList(), template.get("wildcardlist"));
-		Assert.assertEquals(new ArrayList(), template.get("longlist"));
-		Assert.assertEquals("", template.get("namechar"));
-		Assert.assertEquals(new HashMap(), template.get("map"));
-		Assert.assertEquals(new HashMap(), template.get("mapstringinteger"));
-		Assert.assertEquals(new HashMap(), template.get("mapsubobjinteger"));
-		Assert.assertEquals(new HashMap(), template.get("mapintegersubobj"));
-		Assert.assertEquals(new HashMap(), template.get("mapintegerlistsubsubobj"));
+		assertEquals(0, template.get("my_id"));
+		assertEquals(0, template.get("idint"));
+		assertEquals(0, template.get("idlong"));
+		assertEquals("", template.get("name"));
+		assertEquals("", template.get("gender"));
+		assertEquals(true, template.get("bool"));
+		assertEquals(new ArrayList(), template.get("intarrarr"));
+		assertEquals(new JSONDocTemplate(), template.get("sub_obj"));
+		assertEquals(new ArrayList(), template.get("untypedlist"));
+		assertEquals(new ArrayList(), template.get("subsubobjarr"));
+		assertEquals(new ArrayList(), template.get("stringlist"));
+		assertEquals(new ArrayList(), template.get("stringarrarr"));
+		assertEquals(new ArrayList(), template.get("integerarr"));
+		assertEquals(new ArrayList(), template.get("stringarr"));
+		assertEquals(new ArrayList(), template.get("intarr"));
+		assertEquals(new ArrayList(), template.get("subobjlist"));
+		assertEquals(new ArrayList(), template.get("wildcardlist"));
+		assertEquals(new ArrayList(), template.get("longlist"));
+		assertEquals("", template.get("namechar"));
+		assertEquals(new HashMap(), template.get("map"));
+		assertEquals(new HashMap(), template.get("mapstringinteger"));
+		assertEquals(new HashMap(), template.get("mapsubobjinteger"));
+		assertEquals(new HashMap(), template.get("mapintegersubobj"));
+		assertEquals(new HashMap(), template.get("mapintegerlistsubsubobj"));
 		
 		System.out.println(mapper.writeValueAsString(template));
 	}
@@ -59,17 +57,17 @@ public class JSONDocTemplateBuilderTest {
 	@Test
 	public void testTemplateWithConstant() throws Exception {
         final ObjectMapper mapper = new ObjectMapper();
-        final Set<Class<?>> classes = Sets.<Class<?>>newHashSet(ClassWithConstant.class);
+        final Set<Class<?>> classes = Set.of(ClassWithConstant.class);
 
         final Map<String, Object> template = JSONDocTemplateBuilder.build(ClassWithConstant.class, classes);
-        Assert.assertEquals("", template.get("identifier"));
-        Assert.assertEquals(null, template.get(THIS_IS_A_CONSTANT));
+        assertEquals("", template.get("identifier"));
+        assertNull(template.get(THIS_IS_A_CONSTANT));
 
         final String serializedTemplate =
             "{" +
                 "\"identifier\":\"\"" +
             "}";
 
-        assertThat(mapper.writeValueAsString(template), is(serializedTemplate));
+        assertEquals(mapper.writeValueAsString(template), serializedTemplate);
 	}
 }
