@@ -9,8 +9,6 @@ import java.util.UUID;
 
 import org.jsondoc.core.pojo.JSONDoc.MethodDisplay;
 
-import com.google.common.collect.Sets;
-
 public class ApiMethodDoc extends AbstractDoc implements Comparable<ApiMethodDoc> {
 	public final String jsondocId = UUID.randomUUID().toString();
 
@@ -42,19 +40,19 @@ public class ApiMethodDoc extends AbstractDoc implements Comparable<ApiMethodDoc
 		this.id = null;
 		this.description = "";
 		this.summary = "";
-		this.path = new LinkedHashSet<String>();
-		this.verb = new LinkedHashSet<ApiVerb>();
-		this.produces = new LinkedHashSet<String>();
-		this.consumes = new LinkedHashSet<String>();
-		this.headers = new LinkedHashSet<ApiHeaderDoc>();
-		this.pathparameters = new LinkedHashSet<ApiParamDoc>();
-		this.queryparameters = new LinkedHashSet<ApiParamDoc>();
+		this.path = new LinkedHashSet<>();
+		this.verb = new LinkedHashSet<>();
+		this.produces = new LinkedHashSet<>();
+		this.consumes = new LinkedHashSet<>();
+		this.headers = new LinkedHashSet<>();
+		this.pathparameters = new LinkedHashSet<>();
+		this.queryparameters = new LinkedHashSet<>();
 		this.bodyobject = null;
 		this.response = null;
 		this.responsestatuscode = "";
 		this.visibility = ApiVisibility.UNDEFINED;
 		this.stage = ApiStage.UNDEFINED;
-		this.apierrors = new ArrayList<ApiErrorDoc>();
+		this.apierrors = new ArrayList<>();
 		this.supportedversions = null;
 		this.auth = null;
 		this.displayMethodAs = MethodDisplay.URI;
@@ -209,9 +207,9 @@ public class ApiMethodDoc extends AbstractDoc implements Comparable<ApiMethodDoc
 		case URI:
 			return path;
 		case SUMMARY:
-			return Sets.newHashSet(summary);
+			return Set.of(summary);
 		case METHOD:
-			return Sets.newHashSet(method);
+			return Set.of(method);
 		default:
 			return path;
 		}
@@ -256,7 +254,7 @@ public class ApiMethodDoc extends AbstractDoc implements Comparable<ApiMethodDoc
 			return i;
 
 		if (this.queryparameters.size() == o.getQueryparameters().size()) {
-			Set<ApiParamDoc> bothQueryParameters = new HashSet<ApiParamDoc>();
+			Set<ApiParamDoc> bothQueryParameters = new HashSet<>();
 			bothQueryParameters.addAll(this.queryparameters);
 			bothQueryParameters.addAll(o.getQueryparameters());
 			if (bothQueryParameters.size() > this.queryparameters.size()) {
