@@ -1,12 +1,11 @@
 package org.jsondoc.core.pojo.flow;
 
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 import org.jsondoc.core.annotation.flow.ApiFlowStep;
 import org.jsondoc.core.pojo.ApiMethodDoc;
-
-import com.google.common.collect.Sets;
 
 public class ApiFlowStepDoc {
 	public final String jsondocId = UUID.randomUUID().toString();
@@ -27,7 +26,7 @@ public class ApiFlowStepDoc {
 		
 		if(apiFlowStepDoc.getApimethoddoc() == null) {
 			ApiMethodDoc apiMethodDoc = new ApiMethodDoc();
-			apiMethodDoc.setPath(Sets.newHashSet(String.format(ERROR_MISSING_METHOD_ID, annotation.apimethodid())));
+			apiMethodDoc.setPath(Set.of(String.format(ERROR_MISSING_METHOD_ID, annotation.apimethodid())));
 			apiMethodDoc.addJsondocerror(String.format(ERROR_MISSING_METHOD_ID, annotation.apimethodid()));
 			apiMethodDoc.addJsondochint(HINT_MISSING_METHOD_ID);
 			apiFlowStepDoc.setApimethoddoc(apiMethodDoc);
