@@ -4,10 +4,11 @@ import org.jsondoc.core.pojo.JSONDoc;
 import org.jsondoc.core.pojo.JSONDoc.MethodDisplay;
 import org.jsondoc.core.scanner.DefaultJSONDocScanner;
 import org.jsondoc.core.scanner.JSONDocScanner;
-import org.junit.Assert;
-import org.junit.Test;
 
-import com.google.common.collect.Lists;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
 
 //@ApiObject ignored for ParameterizedType return objects
 //https://github.com/fabiomaffioletti/jsondoc/issues/151
@@ -17,10 +18,10 @@ public class Issue151Test {
 	
 	@Test
 	public void testIssue151() {
-		JSONDoc jsonDoc = jsondocScanner.getJSONDoc("", "", Lists.newArrayList("org.jsondoc.core.issues.issue151"), true, MethodDisplay.URI);
-		Assert.assertEquals(2, jsonDoc.getObjects().keySet().size());
-		Assert.assertEquals(1, jsonDoc.getObjects().get("bargroup").size());
-		Assert.assertEquals(1, jsonDoc.getObjects().get("foogroup").size());
+		JSONDoc jsonDoc = jsondocScanner.getJSONDoc("", "", List.of("org.jsondoc.core.issues.issue151"), true, MethodDisplay.URI);
+		Assertions.assertEquals(2, jsonDoc.getObjects().keySet().size());
+		Assertions.assertEquals(1, jsonDoc.getObjects().get("bargroup").size());
+		Assertions.assertEquals(1, jsonDoc.getObjects().get("foogroup").size());
 	}
 
 }
