@@ -5,17 +5,16 @@ import java.util.Map;
 import java.util.Set;
 
 import org.jsondoc.core.util.pojo.MyEnum;
-import org.junit.Test;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Sets;
+import org.junit.jupiter.api.Test;
 
 public class JSONDocEnumTemplateBuilderTest {
 
 	@Test
-	public void testTemplate() throws IOException, IllegalArgumentException, IllegalAccessException, InstantiationException {
+	public void testTemplate() throws IOException, IllegalArgumentException {
 		ObjectMapper mapper = new ObjectMapper();
-		Set<Class<?>> classes = Sets.<Class<?>>newHashSet(MyEnum.class);
+		Set<Class<?>> classes = Set.of(MyEnum.class);
 		
 		Map<String, Object> template = JSONDocTemplateBuilder.build(MyEnum.class, classes);
 		System.out.println(mapper.writeValueAsString(template));
