@@ -5,11 +5,10 @@ import java.util.Set;
 
 import org.jsondoc.core.annotation.ApiObject;
 import org.jsondoc.core.annotation.ApiObjectField;
-import org.junit.Assert;
-import org.junit.Test;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Sets;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class JSONDocTemplateBuilderWithOrderTest {
 
@@ -40,13 +39,13 @@ public class JSONDocTemplateBuilderWithOrderTest {
 	@Test
 	public void thatTemplateIsMappedToStringCorrectly() throws Exception {
 		final ObjectMapper mapper = new ObjectMapper();
-		Set<Class<?>> classes = Sets.<Class<?>>newHashSet(Unordered.class, Ordered.class);
+		Set<Class<?>> classes = Set.of(Unordered.class, Ordered.class);
 
 		Map<String, Object> unorderedTemplate = JSONDocTemplateBuilder.build(Unordered.class, classes);
-		Assert.assertEquals("{\"aField\":\"\",\"xField\":\"\"}", mapper.writeValueAsString(unorderedTemplate));
+		Assertions.assertEquals("{\"aField\":\"\",\"xField\":\"\"}", mapper.writeValueAsString(unorderedTemplate));
 
 		Map<String, Object> orderedTemplate = JSONDocTemplateBuilder.build(Ordered.class, classes);
-		Assert.assertEquals("{\"xField\":\"\",\"aField\":\"\",\"bField\":\"\"}", mapper.writeValueAsString(orderedTemplate));
+		Assertions.assertEquals("{\"xField\":\"\",\"aField\":\"\",\"bField\":\"\"}", mapper.writeValueAsString(orderedTemplate));
 	}
 
 }
