@@ -9,20 +9,17 @@ import org.jsondoc.core.util.pojo.StackOverflowTemplateObjectTwo;
 import org.jsondoc.core.util.pojo.StackOverflowTemplateSelf;
 import org.jsondoc.core.util.pojo.NotAnnotatedStackOverflowObjectOne;
 import org.jsondoc.core.util.pojo.NotAnnotatedStackOverflowObjectTwo;
-import org.junit.Test;
 
-import com.fasterxml.jackson.core.JsonGenerationException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Sets;
+import org.junit.jupiter.api.Test;
 
 public class StackOverflowTemplateBuilderTest {
 
 	private ObjectMapper mapper = new ObjectMapper();
 	
 	@Test
-	public void testTemplate() throws JsonGenerationException, JsonMappingException, IOException, IllegalArgumentException, IllegalAccessException, InstantiationException {
-		Set<Class<?>> classes = Sets.<Class<?>>newHashSet(StackOverflowTemplateSelf.class, StackOverflowTemplateObjectOne.class, StackOverflowTemplateObjectTwo.class);
+	public void testTemplate() throws IOException, IllegalArgumentException {
+		Set<Class<?>> classes = Set.of(StackOverflowTemplateSelf.class, StackOverflowTemplateObjectOne.class, StackOverflowTemplateObjectTwo.class);
 		
 		StackOverflowTemplateSelf objectSelf = new StackOverflowTemplateSelf();
 		Map<String, Object> template = JSONDocTemplateBuilder.build(objectSelf.getClass(), classes);
@@ -38,8 +35,8 @@ public class StackOverflowTemplateBuilderTest {
 	}
 	
 	@Test
-	public void typeOneTwo() throws JsonGenerationException, JsonMappingException, IOException {
-		Set<Class<?>> classes = Sets.<Class<?>>newHashSet(NotAnnotatedStackOverflowObjectOne.class, NotAnnotatedStackOverflowObjectTwo.class);
+	public void typeOneTwo() throws IOException {
+		Set<Class<?>> classes = Set.of(NotAnnotatedStackOverflowObjectOne.class, NotAnnotatedStackOverflowObjectTwo.class);
 		
 		NotAnnotatedStackOverflowObjectOne typeOne = new NotAnnotatedStackOverflowObjectOne();
 		Map<String, Object> template = JSONDocTemplateBuilder.build(typeOne.getClass(), classes);
